probe: extract types, helpers, and ProbeCache to internal/probe
Move every part of the probe subsystem that is already independent of
the Server struct into a new internal/probe package:

  types.go: ProbeRequest, ProbeProviderRequest, ProbeProviderResponseData,
            LightweightProbeRequest, LightweightProbeResponseData,
            ModelProbeRequest, ProbeTarget/ProbeMode + constants,
            ProbeV2Request, ProbeV2Data, ProbeV2ResponseChunk,
            ValidationError, ValidateProbeV2Request, ProbeMessage,
            ScenarioEndpoint
  cache.go: ProbeCache, ModelEndpointCapability, EndpointStatus,
            CachedModelCapability, ProbeResult, ProbeCacheRequest,
            NewProbeCache + methods (TTL accessor added)

Call sites in internal/server now reference probe.X directly — no
transitional aliases, no compatibility shims.

What stayed in internal/server (still coupled to *Server, out of scope
for this round):
  - AdaptiveProbe, LightweightProbe, V2 probe strategies and handlers
  - Response wrapper types ProbeProviderResponse, LightweightProbeResponse,
    ProbeV2Response (embed *ErrorDetail, which is the server's global
    error model used in 100+ places)
  - testProviderConnectivity, InvalidateProviderCache helpers
  - HTTP routing in webui.go, adaptive_endpoint_router.go integration,
    health monitor in server.go

These remain in internal/server because they depend on Server fields
(clientPool, config, smart routing, ErrorDetail) and require real
decoupling work — separate follow-up.

Build/test/vet green:
  go build ./...
  go vet ./internal/probe/... ./internal/server/...
  go test ./internal/server/... ./internal/probe/...
