+++ internal/server/probe_handler.go
@@ -4,13 +4,14 @@ import (
 	"context"
 	"time"
 
+	"github.com/tingly-dev/tingly-box/internal/probe"
 	"github.com/tingly-dev/tingly-box/internal/protocol"
 	"github.com/tingly-dev/tingly-box/internal/typ"
 )
 
 // testProviderConnectivity tests if a provider's API key and connectivity are working.
 // Kept for use by provider_handler.go provider onboarding/verification.
-func (s *Server) testProviderConnectivity(req *ProbeProviderRequest) (bool, string, int, error) {
+func (s *Server) testProviderConnectivity(req *probe.ProbeProviderRequest) (bool, string, int, error) {
 	provider := &typ.Provider{
 		Name:     req.Name,
 		APIBase:  req.APIBase,
@@ -27,7 +28,7 @@ func (s *Server) testProviderConnectivity(req *ProbeProviderRequest) (bool, stri
 	message := "Hello, this is a test message. Please respond with a short greeting."
 
 	// Use SDK-based probe from probe_v2_sdk.go
-	data, err := s.probeProviderWithSDK(ctx, provider, model, message, ProbeV2ModeSimple)
+	data, err := s.probeProviderWithSDK(ctx, provider, model, message, probe.ProbeV2ModeSimple)
 	if err != nil {
 		return false, err.Error(), 0, nil
 	}
